some bug is fixed by gulp injection

+++ src/app/services/Authentication.service.js
@@ -45,7 +45,7 @@ export class AuthenticationService {
             .then(res => {
                 if(res && res.status.code === '0000') {
                     /*@LOG*/ this.$log.debug('MEMBER INFO IS LOADED');
-                    
+
                     this.$rootScope.member = res.result;
                     this.CookieService.put('member', this.$rootScope.member);
                     if(this.$rootScope.member.country) this.AppSettingService.set('country', this.$rootScope.member.country.alpha2Code);
@@ -56,6 +56,7 @@ export class AuthenticationService {
 
                 defer.resolve();
             }, err => {
+                /*@LOG*/ this.$log.debug(err);
                 this.clear('reload');
                 defer.reject('Authentication init Error!');
             });
@@ -115,6 +116,7 @@ export class AuthenticationService {
             }
             else this.clear('reload');
         }, err => {
+            /*LOG*/ this.$log.debug(err);
             this.clear('reload');
         });
 
@@ -136,6 +138,7 @@ export class AuthenticationService {
             }
             else this.clear('reload');
         }, err => {
+            /*LOG*/ this.$log.debug(err);
             this.clear('reload');
         });
     }
@@ -144,20 +147,23 @@ export class AuthenticationService {
         if(this.$rootScope.memberState.sign && this.$rootScope.member) {
             this.APIService.resource('members.signout').put()
             .then(res => {
-                delete this.$rootScope.member;
+                if(res && res.status.code === '0000') {
+                    delete this.$rootScope.member;
 
-                //DESTROY TOKEN AND AUTH DATA
-                this.CookieService.remove('auth');
-                this.$rootScope.memberState.sign = false;
+                    //DESTROY TOKEN AND AUTH DATA
+                    this.CookieService.remove('auth');
+                    this.$rootScope.memberState.sign = false;
 
-                this.CookieService.putEncrypt('memberState', this.$rootScope.memberState);
+                    this.CookieService.putEncrypt('memberState', this.$rootScope.memberState);
 
-                this.AppSettingService.set('country', this.$rootScope.setting.country_code);
+                    this.AppSettingService.set('country', this.$rootScope.setting.country_code);
 
-                this.$state.go('common.default.main');
+                    this.$state.go('common.default.main');
 
-                if(reload === 'reload') this.$window.location.reload();
+                    if(reload === 'reload') this.$window.location.reload();
+                }
             }, err => {
+                /*LOG*/ this.$log.debug(err);
                 this.$log.error('AUTH CLEAR METHOD IS NOT WORKED :: AuthenticationService');
                 return false;
             });
